+++ src/Plex.Api/Models/MediaContainer.cs
@@ -91,8 +91,8 @@ namespace Plex.Api.Models
         public string TranscoderVideoBitrates { get; set; }
         public string TranscoderVideoQualities { get; set; }
         public string TranscoderVideoResolutions { get; set; }
-        public int UpdatedAt { get; set; }
-        public int CreatedAt { get; set; }
+        public long UpdatedAt { get; set; }
+        public long CreatedAt { get; set; }
         public bool Updater { get; set; }
         public string Version { get; set; }
         public bool VoiceSearch { get; set; }
